+++ packages/fmql/src/fmql/qlang/__init__.py
@@ -1,3 +1,0 @@
-from fmql.qlang.compile import compile_query
-
-__all__ = ["compile_query"]